Add files via upload
Datasets
The folder contains the .csv files of all the 15 investigated datasets along with classname (name), seven independent variables (CBO, NOC, LCOM, DIT, SLOC, RFC, WMC) and the binary dependent variable change with values "yes" or "no".

Fitness Variants 30 runs
This folder contains the 30 runs of each of the seven fitness variants for the 15 investigated datasets. The confusion matric statistics (TP, TN, FP and FN) are reported for 30 runs along with the corresponding G-mean1 and Balance values.

ASOF Training Data
This folder contains .csv files of the training data which was collected for each of the 15 datasets after the application of rules mentioned in Section 2 (Figure 4). Each .csv file contains the selected independent variables of the corresponding datasets along with the best fitness variant for the corresponding datapoint.